@@ -20,43 +20,54 @@ from hashlib import sha256
 from io import BytesIO
 from os import urandom
 
-from pyrogram.api.core import Message, Long
-from . import AES, KDF
+from pyrogram.raw.core import Message, Long
+from . import aes
 
 
-class MTProto:
-    @staticmethod
-    def pack(message: Message, salt: int, session_id: bytes, auth_key: bytes, auth_key_id: bytes) -> bytes:
-        data = Long(salt) + session_id + message.write()
-        padding = urandom(-(len(data) + 12) % 16 + 12)
+def kdf(auth_key: bytes, msg_key: bytes, outgoing: bool) -> tuple:
+    # https://core.telegram.org/mtproto/description#defining-aes-key-and-initialization-vector
+    x = 0 if outgoing else 8
 
-        # 88 = 88 + 0 (outgoing message)
-        msg_key_large = sha256(auth_key[88: 88 + 32] + data + padding).digest()
-        msg_key = msg_key_large[8:24]
-        aes_key, aes_iv = KDF(auth_key, msg_key, True)
+    sha256_a = sha256(msg_key + auth_key[x: x + 36]).digest()
+    sha256_b = sha256(auth_key[x + 40:x + 76] + msg_key).digest()  # 76 = 40 + 36
 
-        return auth_key_id + msg_key + AES.ige256_encrypt(data + padding, aes_key, aes_iv)
+    aes_key = sha256_a[:8] + sha256_b[8:24] + sha256_a[24:32]
+    aes_iv = sha256_b[:8] + sha256_a[8:24] + sha256_b[24:32]
 
-    @staticmethod
-    def unpack(b: BytesIO, session_id: bytes, auth_key: bytes, auth_key_id: bytes) -> Message:
-        assert b.read(8) == auth_key_id, b.getvalue()
+    return aes_key, aes_iv
 
-        msg_key = b.read(16)
-        aes_key, aes_iv = KDF(auth_key, msg_key, False)
-        data = BytesIO(AES.ige256_decrypt(b.read(), aes_key, aes_iv))
-        data.read(8)
 
-        # https://core.telegram.org/mtproto/security_guidelines#checking-session-id
-        assert data.read(8) == session_id
+def pack(message: Message, salt: int, session_id: bytes, auth_key: bytes, auth_key_id: bytes) -> bytes:
+    data = Long(salt) + session_id + message.write()
+    padding = urandom(-(len(data) + 12) % 16 + 12)
 
-        message = Message.read(data)
+    # 88 = 88 + 0 (outgoing message)
+    msg_key_large = sha256(auth_key[88: 88 + 32] + data + padding).digest()
+    msg_key = msg_key_large[8:24]
+    aes_key, aes_iv = kdf(auth_key, msg_key, True)
 
-        # https://core.telegram.org/mtproto/security_guidelines#checking-sha256-hash-value-of-msg-key
-        # https://core.telegram.org/mtproto/security_guidelines#checking-message-length
-        # 96 = 88 + 8 (incoming message)
-        assert msg_key == sha256(auth_key[96:96 + 32] + data.getvalue()).digest()[8:24]
+    return auth_key_id + msg_key + aes.ige256_encrypt(data + padding, aes_key, aes_iv)
 
-        # https://core.telegram.org/mtproto/security_guidelines#checking-msg-id
-        assert message.msg_id % 2 != 0
 
-        return message
+def unpack(b: BytesIO, session_id: bytes, auth_key: bytes, auth_key_id: bytes) -> Message:
+    assert b.read(8) == auth_key_id, b.getvalue()
+
+    msg_key = b.read(16)
+    aes_key, aes_iv = kdf(auth_key, msg_key, False)
+    data = BytesIO(aes.ige256_decrypt(b.read(), aes_key, aes_iv))
+    data.read(8)
+
+    # https://core.telegram.org/mtproto/security_guidelines#checking-session-id
+    assert data.read(8) == session_id
+
+    message = Message.read(data)
+
+    # https://core.telegram.org/mtproto/security_guidelines#checking-sha256-hash-value-of-msg-key
+    # https://core.telegram.org/mtproto/security_guidelines#checking-message-length
+    # 96 = 88 + 8 (incoming message)
+    assert msg_key == sha256(auth_key[96:96 + 32] + data.getvalue()).digest()[8:24]
+
+    # https://core.telegram.org/mtproto/security_guidelines#checking-msg-id
+    assert message.msg_id % 2 != 0
+
+    return message
